Fix #10201 and #11682: Editor and translation tutorials fully working (#11857)
* initial commit

* fix tests

* lint

* Remove

* small

* fix test

* attempt to fix 100% cov
