+++ core/templates/pages/exploration-editor-page/editor-tab/exploration-editor-tab.component.ts
@@ -46,35 +46,54 @@ require(
 require('pages/exploration-editor-page/services/graph-data.service.ts');
 require('pages/exploration-editor-page/services/router.service.ts');
 require(
+  'pages/exploration-editor-page/services/' +
+  'state-tutorial-first-time.service.ts');
+require(
   'pages/exploration-editor-page/services/state-editor-refresh.service.ts');
+require(
+  'pages/exploration-editor-page/services/' +
+  'user-exploration-permissions.service.ts');
 require('components/state-editor/state-editor.directive.ts');
 require(
   'components/state-editor/state-editor-properties-services/' +
   'state-editor.service.ts');
 require('services/alerts.service.ts');
 require('services/context.service.ts');
+require('services/editability.service.ts');
 require('services/exploration-features.service.ts');
+require('services/site-analytics.service.ts');
 
 import { Subscription } from 'rxjs';
 
 angular.module('oppia').component('explorationEditorTab', {
   template: require('./exploration-editor-tab.component.html'),
   controller: [
-    '$scope', '$uibModal', 'ExplorationCorrectnessFeedbackService',
-    'ExplorationFeaturesService', 'ExplorationInitStateNameService',
-    'ExplorationStatesService', 'ExplorationWarningsService',
-    'GraphDataService', 'LoaderService', 'RouterService',
-    'StateEditorRefreshService', 'StateEditorService',
-    'UrlInterpolationService',
+    '$scope', '$templateCache', '$uibModal', 'EditabilityService',
+    'ExplorationCorrectnessFeedbackService', 'ExplorationFeaturesService',
+    'ExplorationInitStateNameService', 'ExplorationStatesService',
+    'ExplorationWarningsService', 'GraphDataService', 'LoaderService',
+    'RouterService', 'SiteAnalyticsService', 'StateEditorRefreshService',
+    'StateEditorService', 'StateTutorialFirstTimeService',
+    'UrlInterpolationService', 'UserExplorationPermissionsService',
     function(
-        $scope, $uibModal, ExplorationCorrectnessFeedbackService,
-        ExplorationFeaturesService, ExplorationInitStateNameService,
-        ExplorationStatesService, ExplorationWarningsService,
-        GraphDataService, LoaderService, RouterService,
-        StateEditorRefreshService, StateEditorService,
-        UrlInterpolationService) {
+        $scope, $templateCache, $uibModal, EditabilityService,
+        ExplorationCorrectnessFeedbackService, ExplorationFeaturesService,
+        ExplorationInitStateNameService, ExplorationStatesService,
+        ExplorationWarningsService, GraphDataService, LoaderService,
+        RouterService, SiteAnalyticsService, StateEditorRefreshService,
+        StateEditorService, StateTutorialFirstTimeService,
+        UrlInterpolationService, UserExplorationPermissionsService) {
       var ctrl = this;
       ctrl.directiveSubscriptions = new Subscription();
+      // Replace the ng-joyride template with one that uses <[...]>
+      // interpolators instead of/ {{...}} interpolators.
+      var ngJoyrideTemplate = $templateCache.get(
+        'ng-joyride-title-tplv1.html');
+      ngJoyrideTemplate = ngJoyrideTemplate.replace(
+        /\{\{/g, '<[').replace(/\}\}/g, ']>');
+      $templateCache.put(
+        'ng-joyride-title-tplv1.html', ngJoyrideTemplate);
+
       ctrl.getStateContentPlaceholder = function() {
         if (
           StateEditorService.getActiveStateName() ===
@@ -137,6 +156,9 @@ angular.module('oppia').component('explorationEditorTab', {
           }
 
           LoaderService.hideLoadingScreen();
+        }
+        if (EditabilityService.inTutorialMode()) {
+          ctrl.startTutorial();
         }
       };
 
@@ -287,6 +309,175 @@ angular.module('oppia').component('explorationEditorTab', {
       ctrl.$onDestroy = function() {
         ctrl.directiveSubscriptions.unsubscribe();
       };
+
+      var _ID_TUTORIAL_STATE_CONTENT = '#tutorialStateContent';
+      var _ID_TUTORIAL_STATE_INTERACTION = '#tutorialStateInteraction';
+      var _ID_TUTORIAL_PREVIEW_TAB = '#tutorialPreviewTab';
+      var _ID_TUTORIAL_SAVE_BUTTON = '#tutorialSaveButton';
+
+      var saveButtonTutorialElement = {
+        type: 'element',
+        selector: _ID_TUTORIAL_SAVE_BUTTON,
+        heading: 'Save',
+        text: (
+          'When you\'re done making changes, ' +
+          'be sure to save your work.<br><br>'),
+        placement: 'bottom'
+      };
+
+      ctrl.EDITOR_TUTORIAL_OPTIONS = [{
+        type: 'title',
+        heading: 'Creating in Oppia',
+        text: (
+          'Explorations are learning experiences that you create using ' +
+          'Oppia. Think of explorations as a conversation between a ' +
+          'student and a tutor.')
+      }, {
+        type: 'function',
+        fn: function(isGoingForward) {
+          $('html, body').animate({
+            scrollTop: (isGoingForward ? 0 : 20)
+          }, 1000);
+        }
+      }, {
+        type: 'element',
+        selector: _ID_TUTORIAL_STATE_CONTENT,
+        heading: 'Content',
+        text: (
+          '<p>An Oppia exploration is divided into several \'cards\'. ' +
+          'The first part of a card is the <b>content</b>.</p>' +
+          '<p>Use the content section to set the scene. ' +
+          'Tell the learner a story, give them some information, ' +
+          'and then ask a relevant question.</p>'),
+        placement: 'bottom'
+      }, {
+        type: 'function',
+        fn: function(isGoingForward) {
+          var idToScrollTo = (
+            isGoingForward ? _ID_TUTORIAL_STATE_INTERACTION :
+            _ID_TUTORIAL_STATE_CONTENT);
+          $('html, body').animate({
+            scrollTop: angular.element(idToScrollTo).offset().top - 200
+          }, 1000);
+        }
+      }, {
+        type: 'title',
+        selector: _ID_TUTORIAL_STATE_INTERACTION,
+        heading: 'Interaction',
+        text: (
+          '<p>After you\'ve written the content of your conversation, ' +
+          'choose an <b>interaction type</b>. ' +
+          'An interaction is how you want your learner to respond ' +
+          'to your question.</p> ' +
+          '<p>Oppia has several built-in interactions, including:</p>' +
+          '<ul>' +
+          '  <li>' +
+          '    Multiple Choice' +
+          '  </li>' +
+          '  <li>' +
+          '    Text/Number input' +
+          '  </li>' +
+          '  <li>' +
+          '    Code snippets' +
+          '  </li>' +
+          '</ul>' +
+          'and more.')
+      }, {
+        type: 'function',
+        fn: function(isGoingForward) {
+          var idToScrollTo = (
+            isGoingForward ? _ID_TUTORIAL_PREVIEW_TAB :
+            _ID_TUTORIAL_STATE_INTERACTION);
+          $('html, body').animate({
+            scrollTop: angular.element(idToScrollTo).offset().top - 200
+          }, 1000);
+        }
+      }, {
+        type: 'title',
+        heading: 'Responses',
+        text: (
+          'After the learner uses the interaction you created, it\'s ' +
+          'your turn again to choose how your exploration will respond ' +
+          'to their input. You can send a learner to a new card or ' +
+          'have them repeat the same card, depending on how they answer.')
+      }, {
+        type: 'function',
+        fn: function(isGoingForward) {
+          var idToScrollTo = (
+            isGoingForward ? _ID_TUTORIAL_PREVIEW_TAB :
+            _ID_TUTORIAL_STATE_INTERACTION);
+          $('html, body').animate({
+            scrollTop: angular.element(idToScrollTo).offset().top - 200
+          }, 1000);
+        }
+      }, {
+        type: 'element',
+        selector: _ID_TUTORIAL_PREVIEW_TAB,
+        heading: 'Preview',
+        text: (
+          'At any time, you can click the <b>preview</b> button to ' +
+          'play through your exploration.'),
+        placement: 'bottom'
+      }, saveButtonTutorialElement, {
+        type: 'title',
+        heading: 'Tutorial Complete',
+        text: (
+          '<h2>Now for the fun part...</h2>' +
+          'That\'s the end of the tour! ' +
+          'To finish up, here are some things we suggest: ' +
+          '<ul>' +
+          '  <li>' +
+          '    Create your first card!' +
+          '  </li>' +
+          '  <li>' +
+          '    Preview your exploration.' +
+          '  </li>' +
+          '  <li>' +
+          '    Check out more resources in the ' +
+          '    <a href="https://oppia.github.io/#/" target="_blank">' +
+          '      Help Center.' +
+          '    </a>' +
+          '  </li>' +
+          '</ul>')
+      }];
+
+      ctrl.leaveTutorial = function() {
+        EditabilityService.onEndTutorial();
+        $scope.$apply();
+        StateTutorialFirstTimeService.markEditorTutorialFinished();
+        ctrl.tutorialInProgress = false;
+      };
+
+      ctrl.onSkipTutorial = function() {
+        SiteAnalyticsService.registerSkipTutorialEvent(
+          ctrl.explorationId);
+        ctrl.leaveTutorial();
+      };
+
+      ctrl.onFinishTutorial = function() {
+        SiteAnalyticsService.registerFinishTutorialEvent(
+          ctrl.explorationId);
+        ctrl.leaveTutorial();
+      };
+
+      ctrl.startTutorial = function() {
+        ctrl.tutorialInProgress = true;
+      };
+
+      // Remove save from tutorial if user does not has edit rights for
+      // exploration since in that case Save Draft button will not be
+      // visible on the create page.
+      ctrl.removeTutorialSaveButtonIfNoPermissions = function() {
+        UserExplorationPermissionsService.getPermissionsAsync()
+          .then(function(permissions) {
+            if (!permissions.canEdit) {
+              var index = ctrl.EDITOR_TUTORIAL_OPTIONS.indexOf(
+                saveButtonTutorialElement);
+              ctrl.EDITOR_TUTORIAL_OPTIONS.splice(index, 1);
+            }
+          });
+      };
+      ctrl.removeTutorialSaveButtonIfNoPermissions();
     }
   ]
 });
