+++ core/templates/pages/exploration-editor-page/exploration-editor-page.component.html
@@ -93,7 +93,7 @@
   </div>
 </div>
 <div class="exploration-editor-content" ng-cloak>
-  <div class="container-fluid oppia-editor-page-container" ng-joy-ride="$ctrl.tutorialInProgress" config="$ctrl.EDITOR_TUTORIAL_OPTIONS" on-finish="$ctrl.onFinishTutorial()" on-skip="$ctrl.onSkipTutorial()">
+  <div class="container-fluid oppia-editor-page-container">
     <div class="row" ng-if="$ctrl.ExplorationRightsService.isCloned()">
       <div class="col-lg-12 col-md-12 col-sm-12">
         <div class="oppia-align-center uib-alert alert-warning oppia-alert-warning">
