+++ core/templates/pages/exploration-editor-page/exploration-editor-page.component.ts
@@ -160,7 +160,7 @@ import { Subscription } from 'rxjs';
 angular.module('oppia').component('explorationEditorPage', {
   template: require('./exploration-editor-page.component.html'),
   controller: [
-    '$q', '$rootScope', '$scope', '$templateCache', '$timeout', '$uibModal',
+    '$q', '$rootScope', '$scope', '$uibModal',
     'AutosaveInfoModalsService', 'BottomNavbarStatusService',
     'ChangeListService', 'ContextService',
     'EditabilityService', 'ExplorationAutomaticTextToSpeechService',
@@ -181,7 +181,7 @@ angular.module('oppia').component('explorationEditorPage', {
     'UserEmailPreferencesService', 'UserExplorationPermissionsService',
     'WindowDimensionsService',
     function(
-        $q, $rootScope, $scope, $templateCache, $timeout, $uibModal,
+        $q, $rootScope, $scope, $uibModal,
         AutosaveInfoModalsService, BottomNavbarStatusService,
         ChangeListService, ContextService,
         EditabilityService, ExplorationAutomaticTextToSpeechService,
@@ -204,20 +204,6 @@ angular.module('oppia').component('explorationEditorPage', {
       var ctrl = this;
       ctrl.directiveSubscriptions = new Subscription();
       ctrl.autosaveIsInProgress = false;
-      var _ID_TUTORIAL_STATE_CONTENT = '#tutorialStateContent';
-      var _ID_TUTORIAL_STATE_INTERACTION = '#tutorialStateInteraction';
-      var _ID_TUTORIAL_PREVIEW_TAB = '#tutorialPreviewTab';
-      var _ID_TUTORIAL_SAVE_BUTTON = '#tutorialSaveButton';
-
-      var saveButtonTutorialElement = {
-        type: 'element',
-        selector: _ID_TUTORIAL_SAVE_BUTTON,
-        heading: 'Save',
-        text: (
-          'When you\'re done making changes, ' +
-          'be sure to save your work.<br><br>'),
-        placement: 'bottom'
-      };
 
       var setPageTitle = function() {
         if (ExplorationTitleService.savedMemento) {
@@ -414,33 +400,22 @@ angular.module('oppia').component('explorationEditorPage', {
         return RouterService.getActiveTabName();
       };
 
-      var leaveTutorial = function() {
-        EditabilityService.onEndTutorial();
-        $scope.$apply();
-        StateTutorialFirstTimeService.markEditorTutorialFinished();
-        ctrl.tutorialInProgress = false;
+      ctrl.startEditorTutorial = function() {
+        EditabilityService.onStartTutorial();
+        if (RouterService.getActiveTabName() !== 'main') {
+          ctrl.selectMainTab();
+        } else {
+          StateEditorRefreshService.onRefreshStateEditor.emit();
+        }
       };
 
-      ctrl.onSkipTutorial = function() {
-        SiteAnalyticsService.registerSkipTutorialEvent(ctrl.explorationId);
-        leaveTutorial();
-      };
-
-      ctrl.onFinishTutorial = function() {
-        SiteAnalyticsService.registerFinishTutorialEvent(
-          ctrl.explorationId);
-        leaveTutorial();
-      };
-
-      ctrl.startTutorial = function() {
-        RouterService.navigateToMainTab();
-        // The $timeout wrapper is needed for all components on the page
-        // to load, otherwise elements within ng-if's are not guaranteed to
-        // be present on the page.
-        $timeout(function() {
-          EditabilityService.onStartTutorial();
-          ctrl.tutorialInProgress = true;
-        });
+      ctrl.startTranslationTutorial = function() {
+        EditabilityService.onStartTutorial();
+        if (RouterService.getActiveTabName() !== 'translation') {
+          ctrl.selectTranslationTab();
+        } else {
+          RouterService.onRefreshTranslationTab.emit();
+        }
       };
 
       ctrl.showWelcomeExplorationModal = function() {
@@ -454,7 +429,7 @@ angular.module('oppia').component('explorationEditorPage', {
         }).result.then(function(explorationId) {
           SiteAnalyticsService.registerAcceptTutorialModalEvent(
             explorationId);
-          ctrl.startTutorial();
+          ctrl.startEditorTutorial();
         }, function(explorationId) {
           SiteAnalyticsService.registerDeclineTutorialModalEvent(
             explorationId);
@@ -547,7 +522,13 @@ angular.module('oppia').component('explorationEditorPage', {
         ctrl.directiveSubscriptions.add(
           StateTutorialFirstTimeService.onOpenEditorTutorial.subscribe(
             () => {
-              ctrl.startTutorial();
+              ctrl.startEditorTutorial();
+            })
+        );
+        ctrl.directiveSubscriptions.add(
+          StateTutorialFirstTimeService.onOpenTranslationTutorial.subscribe(
+            () => {
+              ctrl.startTranslationTutorial();
             })
         );
         ctrl.EditabilityService = EditabilityService;
@@ -569,132 +550,6 @@ angular.module('oppia').component('explorationEditorPage', {
         // is also called in $scope.$on when some external events are
         // triggered.
         ctrl.initExplorationPage();
-        ctrl.EDITOR_TUTORIAL_OPTIONS = [{
-          type: 'title',
-          heading: 'Creating in Oppia',
-          text: (
-            'Explorations are learning experiences that you create using ' +
-            'Oppia. Think of explorations as a conversation between a ' +
-            'student and a tutor.')
-        }, {
-          type: 'function',
-          fn: function(isGoingForward) {
-            $('html, body').animate({
-              scrollTop: (isGoingForward ? 0 : 20)
-            }, 1000);
-          }
-        }, {
-          type: 'element',
-          selector: _ID_TUTORIAL_STATE_CONTENT,
-          heading: 'Content',
-          text: (
-            '<p>An Oppia exploration is divided into several \'cards\'. ' +
-            'The first part of a card is the <b>content</b>.</p>' +
-            '<p>Use the content section to set the scene. ' +
-            'Tell the learner a story, give them some information, ' +
-            'and then ask a relevant question.</p>'),
-          placement: 'bottom'
-        }, {
-          type: 'function',
-          fn: function(isGoingForward) {
-            var idToScrollTo = (
-              isGoingForward ? _ID_TUTORIAL_STATE_INTERACTION :
-              _ID_TUTORIAL_STATE_CONTENT);
-            $('html, body').animate({
-              scrollTop: angular.element(idToScrollTo).offset().top - 200
-            }, 1000);
-          }
-        }, {
-          type: 'title',
-          selector: _ID_TUTORIAL_STATE_INTERACTION,
-          heading: 'Interaction',
-          text: (
-            '<p>After you\'ve written the content of your conversation, ' +
-            'choose an <b>interaction type</b>. ' +
-            'An interaction is how you want your learner to respond ' +
-            'to your question.</p> ' +
-            '<p>Oppia has several built-in interactions, including:</p>' +
-            '<ul>' +
-            '  <li>' +
-            '    Multiple Choice' +
-            '  </li>' +
-            '  <li>' +
-            '    Text/Number input' +
-            '  </li>' +
-            '  <li>' +
-            '    Code snippets' +
-            '  </li>' +
-            '</ul>' +
-            'and more.')
-        }, {
-          type: 'function',
-          fn: function(isGoingForward) {
-            var idToScrollTo = (
-              isGoingForward ? _ID_TUTORIAL_PREVIEW_TAB :
-              _ID_TUTORIAL_STATE_INTERACTION);
-            $('html, body').animate({
-              scrollTop: angular.element(idToScrollTo).offset().top - 200
-            }, 1000);
-          }
-        }, {
-          type: 'title',
-          heading: 'Responses',
-          text: (
-            'After the learner uses the interaction you created, it\'s ' +
-            'your turn again to choose how your exploration will respond ' +
-            'to their input. You can send a learner to a new card or ' +
-            'have them repeat the same card, depending on how they answer.')
-        }, {
-          type: 'function',
-          fn: function(isGoingForward) {
-            var idToScrollTo = (
-              isGoingForward ? _ID_TUTORIAL_PREVIEW_TAB :
-              _ID_TUTORIAL_STATE_INTERACTION);
-            $('html, body').animate({
-              scrollTop: angular.element(idToScrollTo).offset().top - 200
-            }, 1000);
-          }
-        }, {
-          type: 'element',
-          selector: _ID_TUTORIAL_PREVIEW_TAB,
-          heading: 'Preview',
-          text: (
-            'At any time, you can click the <b>preview</b> button to ' +
-            'play through your exploration.'),
-          placement: 'bottom'
-        }, saveButtonTutorialElement, {
-          type: 'title',
-          heading: 'Tutorial Complete',
-          text: (
-            '<h2>Now for the fun part...</h2>' +
-            'That\'s the end of the tour! ' +
-            'To finish up, here are some things we suggest: ' +
-            '<ul>' +
-            '  <li>' +
-            '    Create your first card!' +
-            '  </li>' +
-            '  <li>' +
-            '    Preview your exploration.' +
-            '  </li>' +
-            '  <li>' +
-            '    Check out more resources in the ' +
-            '    <a href="https://oppia.github.io/#/" target="_blank">' +
-            '      Help Center.' +
-            '    </a>' +
-            '  </li>' +
-            '</ul>')
-        }];
-        // Remove save from tutorial if user does not has edit rights for
-        // exploration since in that case Save Draft button will not be
-        // visible on the create page.
-        UserExplorationPermissionsService.getPermissionsAsync()
-          .then(function(permissions) {
-            if (!permissions.canEdit) {
-              var index = ctrl.EDITOR_TUTORIAL_OPTIONS.indexOf(
-                saveButtonTutorialElement);
-              ctrl.EDITOR_TUTORIAL_OPTIONS.splice(index, 1);
-            }
-          });
 
         let improvementsTabIsEnabled = false;
         $q.when(ExplorationImprovementsService.isImprovementsTabEnabledAsync())
@@ -702,16 +557,6 @@ angular.module('oppia').component('explorationEditorPage', {
             improvementsTabIsEnabled = improvementsTabIsEnabledResponse;
           });
         ctrl.isImprovementsTabEnabled = () => improvementsTabIsEnabled;
-
-        // Replace the ng-joyride template with one that uses <[...]>
-        // interpolators instead of/ {{...}} interpolators.
-        var ngJoyrideTemplate = $templateCache.get(
-          'ng-joyride-title-tplv1.html');
-        ngJoyrideTemplate = ngJoyrideTemplate.replace(
-          /\{\{/g, '<[').replace(/\}\}/g, ']>');
-        $templateCache.put(
-          'ng-joyride-title-tplv1.html', ngJoyrideTemplate);
-        ctrl.tutorialInProgress = false;
       };
       ctrl.$onDestroy = function() {
         ctrl.directiveSubscriptions.unsubscribe();
