+++ core/templates/pages/exploration-editor-page/translation-tab/translation-tab.component.html
@@ -1,5 +1,5 @@
 <div class="oppia-translation-tab" ng-if="showTranslationTabSubDirectives"
-     ng-joy-ride="translationTutorial"
+     ng-joy-ride="tutorialInProgress"
      config="TRANSLATION_TUTORIAL_OPTIONS"
      on-finish="onFinishTutorial()"
      on-skip="onSkipTutorial()">
