+++ core/templates/pages/exploration-editor-page/translation-tab/translation-tab.component.spec.ts
@@ -74,7 +74,6 @@ describe('Translation tab component', function() {
   var stateTutorialFirstTimeService = null;
   var userExplorationPermissionsService = null;
 
-  var openTranslationTutorialEmitter = new EventEmitter();
   var refreshTranslationTabEmitter = new EventEmitter();
   var enterTranslationForTheFirstTimeEmitter = new EventEmitter();
 
@@ -138,8 +137,6 @@ describe('Translation tab component', function() {
     spyOn(contextService, 'getExplorationId').and.returnValue('exp1');
     spyOn(stateEditorService, 'getActiveStateName').and.returnValue(
       'Introduction');
-    spyOnProperty(stateTutorialFirstTimeService, 'onOpenTranslationTutorial')
-      .and.returnValue(openTranslationTutorialEmitter);
     spyOnProperty(
       stateTutorialFirstTimeService, 'onEnterTranslationForTheFirstTime')
       .and.returnValue(enterTranslationForTheFirstTimeEmitter);
@@ -252,7 +249,7 @@ describe('Translation tab component', function() {
 
       expect($scope.isTranslationTabBusy).toBe(false);
       expect($scope.showTranslationTabSubDirectives).toBe(false);
-      expect($scope.translationTutorial).toBe(false);
+      expect($scope.tutorialInProgress).toBe(false);
     });
 
   it('should load translation tab data when translation tab page is' +
@@ -272,6 +269,56 @@ describe('Translation tab component', function() {
     expect(loaderService.hideLoadingScreen).toHaveBeenCalled();
   });
 
+  it('should start tutorial if in tutorial mode on page load with' +
+    ' permissions', () => {
+    spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
+      .returnValue($q.resolve({
+        canVoiceover: true
+      }));
+    spyOn($scope, 'startTutorial').and.callThrough();
+
+    editabilityService.onStartTutorial();
+    ctrl.$onInit();
+    $scope.$apply();
+    $scope.initTranslationTab();
+    $scope.$apply();
+
+    expect(editabilityService.inTutorialMode()).toBe(true);
+    expect($scope.startTutorial).toHaveBeenCalled();
+    expect($scope.tutorialInProgress).toBe(true);
+  });
+
+  it('should not start tutorial if in tutorial mode on page load but' +
+    ' no permissions', () => {
+    spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
+      .returnValue($q.resolve(null));
+
+    editabilityService.onStartTutorial();
+    ctrl.$onInit();
+    $scope.$apply();
+    $scope.initTranslationTab();
+    $scope.$apply();
+
+    expect(editabilityService.inTutorialMode()).toBe(true);
+    expect($scope.tutorialInProgress).toBe(false);
+  });
+
+  it('should not start tutorial if not in tutorial mode on page load', () => {
+    spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
+      .returnValue($q.resolve({
+        canVoiceover: true
+      }));
+
+    editabilityService.onEndTutorial();
+    ctrl.$onInit();
+    $scope.$apply();
+    $scope.initTranslationTab();
+    $scope.$apply();
+
+    expect(editabilityService.inTutorialMode()).toBe(false);
+    expect($scope.tutorialInProgress).toBe(false);
+  });
+
   it('should finish tutorial on clicking the end tutorial button when' +
     ' it has already started', function() {
     spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
@@ -281,17 +328,18 @@ describe('Translation tab component', function() {
 
     ctrl.$onInit();
     $scope.$apply();
+    editabilityService.onStartTutorial();
+    $scope.$apply();
 
     spyOn(editabilityService, 'onEndTutorial');
     spyOn(stateTutorialFirstTimeService, 'markTranslationTutorialFinished');
-    openTranslationTutorialEmitter.emit();
 
     $scope.onFinishTutorial();
 
     expect(editabilityService.onEndTutorial).toHaveBeenCalled();
     expect(stateTutorialFirstTimeService.markTranslationTutorialFinished)
       .toHaveBeenCalled();
-    expect($scope.translationTutorial).toBe(false);
+    expect($scope.tutorialInProgress).toBe(false);
   });
 
   it('should skip tutorial when the skip tutorial button is clicked',
@@ -303,49 +351,18 @@ describe('Translation tab component', function() {
 
       ctrl.$onInit();
       $scope.$apply();
+      editabilityService.onStartTutorial();
+      $scope.$apply();
 
       spyOn(editabilityService, 'onEndTutorial');
       spyOn(stateTutorialFirstTimeService, 'markTranslationTutorialFinished');
-      openTranslationTutorialEmitter.emit();
 
       $scope.onSkipTutorial();
 
       expect(editabilityService.onEndTutorial).toHaveBeenCalled();
       expect(stateTutorialFirstTimeService.markTranslationTutorialFinished)
         .toHaveBeenCalled();
-      expect($scope.translationTutorial).toBe(false);
-    });
-
-  it('should start tutorial when translation tutorial modal is closed',
-    function() {
-      spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
-        .returnValue($q.resolve({
-          canVoiceover: true
-        }));
-
-      ctrl.$onInit();
-      $scope.$apply();
-
-      spyOn(editabilityService, 'onStartTutorial');
-      openTranslationTutorialEmitter.emit();
-
-      expect($scope.translationTutorial).toBe(true);
-      expect(editabilityService.onStartTutorial).toHaveBeenCalled();
-    });
-
-  it('should not start tutorial when user has no permissions',
-    function() {
-      spyOn(userExplorationPermissionsService, 'getPermissionsAsync').and
-        .returnValue($q.resolve(null));
-
-      ctrl.$onInit();
-      $scope.$apply();
-
-      spyOn(editabilityService, 'onStartTutorial');
-      openTranslationTutorialEmitter.emit();
-
-      expect($scope.translationTutorial).toBe(false);
-      expect(editabilityService.onStartTutorial).not.toHaveBeenCalled();
+      expect($scope.tutorialInProgress).toBe(false);
     });
 
   it('should start tutorial when welcome translation modal is closed',
@@ -358,7 +375,6 @@ describe('Translation tab component', function() {
       ctrl.$onInit();
       $scope.$apply();
 
-      spyOn(editabilityService, 'onStartTutorial');
       spyOn(siteAnalyticsService, 'registerAcceptTutorialModalEvent');
       spyOn($uibModal, 'open').and.returnValue({
         result: $q.resolve('exp1')
@@ -366,7 +382,6 @@ describe('Translation tab component', function() {
       enterTranslationForTheFirstTimeEmitter.emit();
       $scope.$apply();
 
-      expect(editabilityService.onStartTutorial).toHaveBeenCalled();
       expect(siteAnalyticsService.registerAcceptTutorialModalEvent)
         .toHaveBeenCalled();
     });
