+++ core/templates/services/editability.service.spec.ts
@@ -54,4 +54,11 @@ describe('EditabilityService', () => {
     editabilityService.markNotEditable();
     expect(editabilityService.isEditable()).toBe(false);
   });
+
+  it('should return correctly when checking if in tutorial mode', () => {
+    editabilityService.onEndTutorial();
+    expect(editabilityService.inTutorialMode()).toBe(false);
+    editabilityService.onStartTutorial();
+    expect(editabilityService.inTutorialMode()).toBe(true);
+  });
 });
